third_party/WebKit/LayoutTests/http/tests/inspector/network/network-worker-fetch-blocked.html

<html>
<head>
<script src="../inspector-test.js"></script>
<script src="../network-test.js"></script>
<script>
function test()
{
    NetworkTestRunner.recordNetwork();
    SDK.multitargetNetworkManager.setBlockingEnabled(true);
    SDK.multitargetNetworkManager.setBlockedPatterns([{url: 'resources/resource.php', enabled: true}]);

    NetworkTestRunner.makeFetchInWorker('resource.php', {}, fetchCallback);

    function fetchCallback(result)
    {
        TestRunner.addResult('Fetch in worker result: ' + result.value);

        var requests = NetworkTestRunner.networkRequests();
        requests.forEach((request) => {
          TestRunner.addResult(request.url());
          TestRunner.addResult("resource.type: " + request.resourceType());
          TestRunner.addResult("request.failed: " + !!request.failed);
        });

        TestRunner.completeTest();
    }
}
</script>
</head>
<body onload="runTest()">
<p>Tests blocking fetch in worker.</p>
</body>
</html>
